+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * Remove WP Rocket CDN ad
 * Remove WP Rocket preloading notice
 * Remove WP Rocket sidebar
+* Reorder Rank Math invasive admin menu
 
 ### 1.3.1: December 29th, 2020
 * Fix Expiration Date had wrong timezone offset